Change offset tables from ints to shorts to save space.

+++ fs/xfs/xfs_alloc_btree.c
@@ -1041,7 +1041,7 @@ xfs_alloc_log_block(
 {
 	int			first;	/* first byte offset logged */
 	int			last;	/* last byte offset logged */
-	static const int	offsets[] = {	/* table of offsets */
+	static const short	offsets[] = {	/* table of offsets */
 		offsetof(xfs_alloc_block_t, bb_magic),
 		offsetof(xfs_alloc_block_t, bb_level),
 		offsetof(xfs_alloc_block_t, bb_numrecs),
